Melhorias na tela de cadastro e edição dos usuários administradores.

+++ Controllers/UsuariosController.cs
@@ -72,6 +72,12 @@ namespace SistemaDeControleDeTCCs.Controllers
             {
                 usuarios = query.ToList();
             }
+            if (TempData["erro_del"] != null && TempData["erro_del"].ToString().Length > 0) 
+            {
+                ModelState.AddModelError(String.Empty,
+                    TempData["erro_del"].ToString());
+            }
+           
 
             return View(usuarios.OrderBy(x => x.Nome));
         }
@@ -80,12 +86,19 @@ namespace SistemaDeControleDeTCCs.Controllers
         public IActionResult AddOrEdit(string id)
         {
 
-            var  tiposUsuarios = _context.TipoUsuario
-                .OrderBy(x => x.DescTipo)
-                .Where(x => x.DescTipo.Contains("Aluno") || x.DescTipo.Contains("Professor"))
-                .ToList();
-           
-            
+            List<TipoUsuario> tiposUsuarios = null;
+            if (User.IsInRole("Administrador"))
+            {
+                tiposUsuarios = _context.TipoUsuario.OrderBy(x => x.DescTipo).Where(x => x.DescTipo.Contains("Aluno")
+                || x.DescTipo.Contains("Professor")
+                || x.DescTipo.Contains("Administrador")).ToList();
+            }
+            else
+            {
+                tiposUsuarios = _context.TipoUsuario.OrderBy(x => x.DescTipo).Where(x => x.DescTipo.Contains("Aluno") || x.DescTipo.Contains("Professor")).ToList();
+
+            }
+
             var usuario = new Usuario();
             if (id != null)
             {
@@ -135,6 +148,14 @@ namespace SistemaDeControleDeTCCs.Controllers
                     {
                         ModelState.AddModelError(String.Empty,
                             "O usuário que você está tentando alterar é um usuário administrador e apenas outro administrador pode realizar esta modificação");
+
+                        return AddOrEdit(userTemp.Id);
+                    }
+                    if (userTemp.TipoUsuarioId == admId
+                       && usuario.TipoUsuarioId != admId)
+                    {
+                        ModelState.AddModelError(String.Empty,
+                            "Não é possível modificar o tipo de usuários que são administradores");
 
                         return AddOrEdit(userTemp.Id);
                     }
@@ -207,7 +228,19 @@ namespace SistemaDeControleDeTCCs.Controllers
         // GET: Usuarios/Delete/5
         public async Task<IActionResult> Delete(string id)
         {
-            var usuario = await _context.Users.FindAsync(id);
+            
+            var usuario =  _context.Users.Include(u => u.TipoUsuario).Where(u => u.Id == id).First();
+            if(usuario == null)
+            {
+                return NotFound();
+            }
+            if (usuario.TipoUsuario.DescTipo.Equals("Administrador"))
+            {
+                
+                TempData["erro_del"] = "O usuário que você está tentando alterar é um usuário administrador e não é permitida essa transação.";
+                return RedirectToAction(nameof(Index));
+            }
+
             _context.Users.Remove(usuario);
 
             _context.LogAuditoria.Add(
